% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPvalueCutoff.R
\name{findPvalueCutoff}
\alias{findPvalueCutoff}
\title{Estimates false positive rate at a series of p-value cutoffs}
\usage{
findPvalueCutoff(
  x = NULL,
  thresholds = NULL,
  range = c(0, 1),
  interval = 0.05,
  nsims = 1000
)
}
\arguments{
\item{x}{A vector of p-values.}

\item{thresholds}{A vector of manually selected p-value cutoffs at which false positive rate needs to be estimated.}

\item{range}{Limit of lower and upper p-value cutoffs within which false positive rate will be estimated. Default when thresholds parameter is NULL.}

\item{interval}{Interval used to generate a series of p-values within lower and upper bound given in range porameter.}

\item{nsims}{Number of simulated (randomized) datasets. Default is 1000 but at least 10000 datasets are recommended.}
}
\value{
A named (p-value cutoffs) vector with estimated false positive rate.
}
\description{
Given a p-value vector, it estimates number of rejected null hypothesis at a series of p-value cutoffs
}
\examples{
\dontrun{
findPvalueCutoff(x = pval, thresholds=c(0.0001, 0.001, 0.01, 0.1, 0.2), nsims = 1000)
findPvalueCutoff(x = pval, range= c(0,2), interval= 0.01, nsims = 1000)
}
}
